Add files via upload
first run ChatServer java class. ==> java ChatServer 8989
When the server starts then go for ChatClient class which is in another directory.
java ChatClient localhost 8989
Run ChatClient from multiple cmd so that you can have multiple user for your application.

This application can work on any system provided it is connected to the same network because for running ChatClient class we are passing localhost as an argument.